python/requirements.txt
requests>=2.31.0
numpy>=1.22.1
stashapp-tools>=0.2.42
opencv-python>=4.0.0.0
ffmpeg-python>=0.2.0